prompt: What is John working on and why?
John is currently working on an open community art project. The project's goal
is to create code that can write and maintain itself based on a standard
description of architecture, a universal blueprint. This blueprint is intended
to be intuitive enough that anyone can begin using it quickly after the correct
way of communicating to that individual is established. The end goal of this
project is to create a software architect named Alice. Alice will be able to
consult on existing projects to provide analysis of their architectures and
properties such as maintainability. She will also provide guidance for developer
workstream prioritization to achieve strategic business goals, and write and
maintain codebases herself, accepting contributions from external contributors.

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>